Improve integration test file structure
